+++ studio/src/app/utils/editor/parse-elements.utils.tsx
@@ -47,10 +47,6 @@ export class ParseElementsUtils {
                 attributes['contenteditable'] = true;
             }
 
-            if (element.nodeName && element.nodeName.toLowerCase() === 'deckgo-lazy-img') {
-                attributes['contenteditable'] = 'false';
-            }
-
             resolve(<Elem {...attributes}>{content}</Elem>);
         });
     }
@@ -72,7 +68,8 @@ export class ParseElementsUtils {
         return attributes.slot &&
             attributes.slot !== 'background' &&
             (!element.nodeName || (element.nodeName.toLowerCase() !== 'code' && element.nodeName.toLowerCase() !== SlotType.CODE)) &&
-            (!element.nodeName || (element.nodeName.toLowerCase() !== 'deckgo-social' && element.nodeName.toLowerCase() !== SlotType.SOCIAL));
+            (!element.nodeName || (element.nodeName.toLowerCase() !== 'deckgo-social' && element.nodeName.toLowerCase() !== SlotType.SOCIAL)) &&
+            (!element.nodeName || (element.nodeName.toLowerCase() !== 'deckgo-lazy-img' && element.nodeName.toLowerCase() !== SlotType.IMG));
     }
 
 }