+++ webcomponents/core/src/components.d.ts
@@ -101,12 +101,16 @@ export namespace Components {
   interface DeckgoSlideCountdown {
     'afterSwipe': () => Promise<void>;
     'beforeSwipe': (_enter: boolean) => Promise<boolean>;
+    'customActions': boolean;
+    'customBackground': boolean;
+    'days': number;
     'hours': number;
     'lazyLoadContent': () => Promise<void>;
     'minutes': number;
     'seconds': number;
     'start': () => Promise<void>;
     'stop': () => Promise<void>;
+    'until': string;
   }
   interface DeckgoSlideGif {
     'afterSwipe': () => Promise<void>;
@@ -363,10 +367,14 @@ declare namespace LocalJSX {
     'revealShowFirst'?: boolean;
   }
   interface DeckgoSlideCountdown extends JSXBase.HTMLAttributes<HTMLDeckgoSlideCountdownElement> {
+    'customActions'?: boolean;
+    'customBackground'?: boolean;
+    'days'?: number;
     'hours'?: number;
     'minutes'?: number;
     'onSlideDidLoad'?: (event: CustomEvent<void>) => void;
     'seconds'?: number;
+    'until'?: string;
   }
   interface DeckgoSlideGif extends JSXBase.HTMLAttributes<HTMLDeckgoSlideGifElement> {
     'alt'?: string;