+++ webcomponents/inline-editor/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="1.0.0-alpha.12"></a>
+# 1.0.0-alpha.12 (2019-08-06)
+
+* images will not be considered, per default, as editable anymore. to turn this feature on, use the property `img-editable`
+
 <a name="1.0.0-alpha.11"></a>
 # 1.0.0-alpha.11 (2019-07-30)